# config/openapi.yaml
openapi: 3.0.3
info:
  title: ZBI Controller - OpenAPI 3.0
  description: |-
    This is
  termsOfService: https://www.zbitech.net/terms
  contact:
    email: info@zbitech.net
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1
...
servers:
  - url: https://apps.zbitech.local/api/v1
...
paths:
...
  /instances:
    post:
      tags:
        - instance
      summary: Create a new zbi instance
      description: Create a new zbi instance
      operationId: createInstance
      requestBody:
        description: Create a new zbi instance
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                project:
                  type: string
                type:
                  type: string
                  description: type of instance
                  enum:
                    - zcash
                    - lwd
                    - zebra
                request:
                  $ref: '#/components/schemas/ResourceRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
  /instances/{projectName}/{instanceName}:
    get:
      tags:
        - instance
      summary: Retrieve a zbi instance
      description: Retrieve a zbi instance
      operationId: retrieveInstance
      parameters:
        - name: projectName
          in: path
          description: project of target instance
          required: true
          explode: true
          schema:
            type: string
        - name: instanceName
          in: path
          description: name of target instance
          required: true
          explode: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
    delete:
      tags:
        - instance
      summary: Delete a zbi instance
      description: Delete a zbi instance
      operationId: deleteInstance
      parameters:
        - name: projectName
          in: path
          description: project of target instance
          required: true
          explode: true
          schema:
            type: string
        - name: instanceName
          in: path
          description: name of target instance
          required: true
          explode: true
          schema:
            type: string
      responses:
        '204':
          description: Instance deleted
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
  /instances/{projectName}/{instanceName}/stop:
    post:
      tags:
        - instance
      summary: Stop a zbi instance
      description: Stop a zbi instance
      operationId: stopInstance
      parameters:
        - name: projectName
          in: path
          description: project of target instance
          required: true
          explode: true
          schema:
            type: string
        - name: instanceName
          in: path
          description: name of target instance
          required: true
          explode: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
  /instances/{projectName}/{instanceName}/start:
    post:
      tags:
        - instance
      summary: Start a zbi instance
      description: Start a zbi instance
      operationId: startInstance
      parameters:
        - name: projectName
          in: path
          description: project of target instance
          required: true
          explode: true
          schema:
            type: string
        - name: instanceName
          in: path
          description: name of target instance
          required: true
          explode: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
...
  /instances/{projectName}/{instanceName}/repair:
    post:
      tags:
        - instance
      summary: Repair a zbi instance
      description: Repair a zbi instance
      operationId: repairInstance
      parameters:
        - name: projectName
          in: path
          description: project of target instance
          required: true
          explode: true
          schema:
            type: string
        - name: instanceName
          in: path
          description: name of target instance
          required: true
          explode: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instance'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized request
        '403':
          description: Forbidden
        '429':
          description: Too many requests
        '500':
          description: Internal Server Error
        '503':
          description: Service Unavailable
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    Instance:
      type: object
      properties:
        id:
          type: string
          description: Instance identifier
          example: 83ffec9b-ec1e-494e-8973-ff43171c2c23
        project:
          type: string
          description: name of instance's project
        type:
          type: string
          description: Type of instance
          example: zcash
          enum:
            - zcash
            - lwd
            - zebra
        name:
          type: string
          description: name of instance
          example: main
        status:
          type: string
          description: Status of instance
          enum:
            - new
            - created
            - running
            - stopped
            - pending
            - progressing
            - failed
            - terminating
            - deleted
    ResourceRequest:
      type: object
      properties:
        volumeType:
          type: string
          description: Type of volume
          enum:
            - eph
            - pvc
        volumeSize:
          type: string
          description: Size of volume
          example: 1Gi
        sourceType:
          type: string
          description: Source of volume
          enum:
            - none
            - new
            - volume
            - snapshot
        sourceName:
          type: string
          description: name of source
        sourceProject:
          type: string
          description: project for the volume source
        cpu:
          type: string
          description: cpu request for the instance
          example: 1Gi
        memory:
          type: string
          description: memory request for the instance
        peers:
          type: array
          description: list of peer instances
          items:
            type: string
...
    FieldError:
      type: object
      properties:
        name:
          type: string
          description: name of field
        code:
          type: string
          description: error code
        error:
          type: string
          description: error message
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
        fields:
          type: array
          items:
            $ref: '#/components/schemas/ArrayOfFieldErrors'
...
    ArrayOfFieldErrors:
      type: array
      items:
        $ref: '#/components/schemas/FieldError'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT